Bug 40103 - various LDAP enhancements
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/branches/rel-2-2@516970 13f79535-47bb-0310-9956-ffa450edef68

+++ src/protocol/ldap/org/apache/jmeter/protocol/ldap/sampler/LdapExtClient.java
@@ -67,15 +67,31 @@ public class LdapExtClient {
 	 * @exception NamingException
 	 *                Description of Exception
 	 */
-	public DirContext connect(String host, String port, String rootdn, String username, String password)
+	public DirContext connect(String host, String port, String rootdn, String username, String password, String connTimeOut, boolean secure)
 			throws NamingException {
 		DirContext dirContext;
 		Hashtable env = new Hashtable();
-		env.put(Context.INITIAL_CONTEXT_FACTORY, "com.sun.jndi.ldap.LdapCtxFactory");
-		env.put(Context.PROVIDER_URL, "ldap://" + host + ":" + port + "/" + rootdn);
-		log.info("prov_url= " + env.get(Context.PROVIDER_URL));
-		env.put(Context.REFERRAL, "throw");
-		env.put("java.naming.batchsize", "0");
+		env.put(Context.INITIAL_CONTEXT_FACTORY, "com.sun.jndi.ldap.LdapCtxFactory"); // $NON-NLS-1$
+		StringBuffer sb = new StringBuffer(80);
+		if (secure) {
+			sb.append("ldaps://"); // $NON-NLS-1$
+		} else {
+			sb.append("ldap://"); // $NON-NLS-1$
+		}
+		sb.append(host);
+		if (port.length()>0){
+			sb.append(":"); // $NON-NLS-1$
+			sb.append(port);
+		}
+		sb.append("/"); // $NON-NLS-1$
+		sb.append(rootdn);
+		env.put(Context.PROVIDER_URL,sb.toString());
+		log.info("prov_url= " + env.get(Context.PROVIDER_URL)); // $NON-NLS-1$
+		if (connTimeOut.length()> 0) {
+            env.put("com.sun.jndi.ldap.connect.timeout", connTimeOut); // $NON-NLS-1$
+	    }
+		env.put(Context.REFERRAL, "throw"); // $NON-NLS-1$
+		env.put("java.naming.batchsize", "0"); // $NON-NLS-1$  // $NON-NLS-2$
 		env.put(Context.SECURITY_CREDENTIALS, password);
 		env.put(Context.SECURITY_PRINCIPAL, username);
 		dirContext = new InitialDirContext(env);
@@ -108,6 +124,9 @@ public class LdapExtClient {
 	 **************************************************************************/
 	public NamingEnumeration searchTest(DirContext dirContext, String searchBase, String searchFilter, int scope, long countlim,
 			int timelim, String[] attrs, boolean retobj, boolean deref) throws NamingException {
+		if (dirContext == null) {
+			throw new NamingException("Context is null");
+		}
 		SearchControls searchcontrols = null;
 		searchcontrols = new SearchControls(scope, countlim, timelim, attrs, retobj, deref);
 		log.debug("scope, countlim, timelim, attrs, retobj, deref= " + searchFilter + scope + countlim + timelim
@@ -124,6 +143,9 @@ public class LdapExtClient {
 	 *            filter filter this value from the base
 	 **************************************************************************/
 	public NamingEnumeration compare(DirContext dirContext, String filter, String entrydn) throws NamingException {
+		if (dirContext == null) {
+			throw new NamingException("Context is null");
+		}
 		SearchControls searchcontrols = new SearchControls(0, 1, 0, new String[0], false, false);
 		return dirContext.search(entrydn, filter, searchcontrols);
 	}
@@ -138,6 +160,9 @@ public class LdapExtClient {
 	 **************************************************************************/
 	public void moddnOp(DirContext dirContext, String ddn, String newdn) throws NamingException {
 		log.debug("ddn and newDn= " + ddn + "@@@@" + newdn);
+		if (dirContext == null) {
+			throw new NamingException("Context is null");
+		}
 		dirContext.rename(ddn, newdn);
 	}
 
@@ -150,6 +175,9 @@ public class LdapExtClient {
 	 *            The string (dn) value
 	 **************************************************************************/
 	public void modifyTest(DirContext dirContext, ModificationItem[] mods, String string) throws NamingException {
+		if (dirContext == null) {
+			throw new NamingException("Context is null");
+		}
 		dirContext.modifyAttributes(string, mods);
 
 	}
@@ -164,6 +192,9 @@ public class LdapExtClient {
 	 **************************************************************************/
     public DirContext createTest(DirContext dirContext, Attributes attributes, String string)
 			throws NamingException {
+		if (dirContext == null) {
+			throw new NamingException("Context is null");
+		}
 		return dirContext.createSubcontext(string, attributes);
 	}
 
@@ -174,6 +205,9 @@ public class LdapExtClient {
 	 *            The string (dn) value
 	 **************************************************************************/
 	public void deleteTest(DirContext dirContext, String string) throws NamingException {
+		if (dirContext == null) {
+			throw new NamingException("Context is null");
+		}
 		dirContext.destroySubcontext(string);
 	}
 }
